# src/backend/gohandlers/go/model_list.go
/*
 * Share my Cart
 *
 * Stay home. Stay safe. I'll bring your groceries along as I get mine.
 *
 * API version: 0.1.0
 * Contact: sharemycart@beimir.net
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

// List is an struct that holds the list items in a slice
type List struct {
	Id string `json:"id,omitempty"`

	Name string `json:"name"`

	Type_ string `json:"type"`
	// indicates that a shopping list or a need list is the default list. This is particularly important when creating a needed item for a shopping list:  The needed item will be created in the default need list.
	IsDefault bool `json:"isDefault,omitempty"`

	Items []ListItem `json:"items,omitempty"`
}