# src/backend/gohandlers/go/model_login.go
// Package swagger ...
/*
 * Share my Cart
 *
 * Stay home. Stay safe. I'll bring your groceries along as I get mine.
 *
 * API version: 0.1.0
 * Contact: sharemycart@beimir.net
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

// Login holds the username and password
type Login struct {
	Username string `json:"username"`

	Password string `json:"password"`
}