# src/backend/gohandlers/go/model_password_reset.go
// Package swagger ...
/*
 * Share my Cart
 *
 * Stay home. Stay safe. I'll bring your groceries along as I get mine.
 *
 * API version: 0.1.0
 * Contact: sharemycart@beimir.net
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

// PasswordReset holds the Password and the PasswordConfirm entities
type PasswordReset struct {
	Password string `json:"password"`

	PasswordConfirm string `json:"password_confirm"`
}